+++ History.md
@@ -1,6 +1,11 @@
 # Version 2.8.0
 Release date: Unreleased
 
+### Fixed
+
+### Added
+* 'check', 'uncheck', and 'choose' will now click the associated label if the checkbox/radio button is not visible
+
 # Version 2.7.0
 Release date: 2016-04-07